Separate progress into its own app
