+++ backend/learning/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from django.conf import settings
 
 
 class Grade(models.Model):
@@ -40,23 +39,3 @@ class Lesson(models.Model):
 
     def __str__(self):
         return f"{self.title} (Course: {self.course.name})"
-
-
-class LessonProgress(models.Model):
-    is_completed = models.BooleanField(default=False)
-    last_accessed = models.DateTimeField(auto_now=True)
-
-    # relations (FKs)
-    user = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        on_delete=models.CASCADE,
-        related_name='lesson_progress'
-    )
-    lesson = models.ForeignKey(
-        Lesson, on_delete=models.CASCADE, related_name='progress')
-
-    class Meta:
-        unique_together = ('user', 'lesson')
-
-    def __str__(self):
-        return f"{self.user} - {self.lesson}: ({'Completed' if self.is_completed else 'In Progress'})"
